+++ extensions/prism-launcher/src/utils/preferences.ts
@@ -1,0 +1,10 @@
+import { type Application, getPreferenceValues } from "@raycast/api";
+
+type Preferences = {
+  installPath: Application;
+  instancesPath: string;
+};
+
+export function getPreferences<T extends keyof Preferences>(key: T): Preferences[T] {
+  return getPreferenceValues<Preferences>()[key];
+}