+++ extensions/raycast-ollama/src/lib/ui/ChatView/form/Model.tsx
@@ -35,9 +35,6 @@ interface FormData {
   serverTools: string;
   modelTools: string;
   keepAliveTools: string;
-  serverEmbedding: string;
-  modelEmbedding: string;
-  keepAliveEmbedding: string;
   mcp_server: string[];
 }
 
@@ -75,16 +72,6 @@ export function FormModel(props: props): JSX.Element {
         if (models.filter((model) => model.name === props.Chat?.models.tools?.tag).length > 0)
           setValue("modelTools", props.Chat.models.tools.tag);
       }
-
-      if (props.Chat.models.embedding && data.has(props.Chat.models.embedding.server_name)) {
-        setValue("serverEmbedding", props.Chat.models.embedding.server_name);
-        const models = (data.get(props.Chat.models.embedding.server_name) as UiModelDetails[]).filter(
-          (model) =>
-            model.capabilities && model.capabilities.findIndex((c) => c === OllamaApiModelCapability.EMBEDDING) !== -1
-        );
-        if (models.filter((model) => model.name === props.Chat?.models.embedding?.tag).length > 0)
-          setValue("modelEmbedding", props.Chat.models.embedding.tag);
-      }
     },
   });
   const { value: McpServer, isLoading: isLoadingMcpServer } = useLocalStorage<McpServerConfig>("mcp_server_config", {
@@ -102,12 +89,6 @@ export function FormModel(props: props): JSX.Element {
   );
   const [CheckboxToolsAdvanced, SetCheckboxToolsAdvanced]: [boolean, React.Dispatch<React.SetStateAction<boolean>>] =
     React.useState(props.Chat?.models.tools && props.Chat.models.tools.keep_alive ? true : false);
-  const [CheckboxEmbedding, SetCheckboxEmbedding]: [boolean, React.Dispatch<React.SetStateAction<boolean>>] =
-    React.useState(props.Chat?.models.embedding ? true : false);
-  const [CheckboxEmbeddingAdvanced, SetCheckboxEmbeddingAdvanced]: [
-    boolean,
-    React.Dispatch<React.SetStateAction<boolean>>
-  ] = React.useState(props.Chat?.models.embedding && props.Chat.models.embedding.keep_alive ? true : false);
   const { handleSubmit, itemProps, setValue } = useForm<FormData>({
     onSubmit(values) {
       Submit(values);
@@ -116,7 +97,6 @@ export function FormModel(props: props): JSX.Element {
       keepAliveMain: props.Chat?.models.main.keep_alive ? props.Chat.models.main.keep_alive : "5m",
       keepAliveVision: props.Chat?.models.vision?.keep_alive ? props.Chat.models.vision.keep_alive : "5m",
       keepAliveTools: props.Chat?.models.tools?.keep_alive ? props.Chat.models.tools.keep_alive : "5m",
-      keepAliveEmbedding: props.Chat?.models.embedding?.keep_alive ? props.Chat.models.embedding.keep_alive : "5m",
     },
     validation: {
       serverMain: FormValidation.Required,
@@ -128,15 +108,8 @@ export function FormModel(props: props): JSX.Element {
       serverTools: CheckboxTools ? FormValidation.Required : undefined,
       modelTools: CheckboxTools ? FormValidation.Required : undefined,
       keepAliveTools: (value) => ValidationKeepAlive(CheckboxToolsAdvanced, value),
-      serverEmbedding: CheckboxEmbedding ? FormValidation.Required : undefined,
-      modelEmbedding: CheckboxEmbedding ? FormValidation.Required : undefined,
-      keepAliveEmbedding: (value) => ValidationKeepAlive(CheckboxEmbeddingAdvanced, value),
     },
   });
-
-  React.useEffect(() => {
-    if (!CheckboxEmbedding) SetCheckboxEmbeddingAdvanced(false);
-  }, [CheckboxEmbedding]);
 
   React.useEffect(() => {
     if (!CheckboxVision) SetCheckboxVisionAdvanced(false);
@@ -157,7 +130,6 @@ export function FormModel(props: props): JSX.Element {
   const InfoModel = "Ollama Model.";
   const InfoVisionCheckbox = "Use a different model for vision when you multimodal capabilities is required.";
   const InfoToolsCheckbox = "Use a different model for tools when tool calling is required.";
-  const InfoEmbeddingCheckbox = "Use a different model for embedding when you want to add a large file in context.";
   const InfoMcpServer =
     'Server can be configured with "Manage Mcp Server" Command. A model with tools capabilities is required.';
 
@@ -175,8 +147,6 @@ export function FormModel(props: props): JSX.Element {
       OllamaServer.set(values.serverVision, await GetOllamaServerByName(values.serverVision));
     if (values.serverTools && !OllamaServer.has(values.serverTools))
       OllamaServer.set(values.serverTools, await GetOllamaServerByName(values.serverTools));
-    if (values.serverEmbedding && !OllamaServer.has(values.serverEmbedding))
-      OllamaServer.set(values.serverEmbedding, await GetOllamaServerByName(values.serverEmbedding));
     let chat = props.Chat;
     if (chat && props.ChatNameIndex !== undefined) {
       chat.models.main = {
@@ -204,16 +174,6 @@ export function FormModel(props: props): JSX.Element {
         };
       } else if (!CheckboxTools && chat.models.tools) {
         chat.models.tools = undefined;
-      }
-      if (values.serverEmbedding && values.modelEmbedding) {
-        chat.models.embedding = {
-          server_name: values.serverEmbedding,
-          server: OllamaServer.get(values.serverEmbedding) as OllamaServer,
-          tag: values.modelEmbedding,
-          keep_alive: CheckboxEmbeddingAdvanced ? values.keepAliveEmbedding : undefined,
-        };
-      } else if (!CheckboxEmbedding && chat.models.embedding) {
-        chat.models.embedding = undefined;
       }
       if (values.mcp_server.length > 0) {
         chat.mcp_server = values.mcp_server;
@@ -254,15 +214,6 @@ export function FormModel(props: props): JSX.Element {
                   server: OllamaServer.get(values.serverTools) as OllamaServer,
                   tag: values.modelTools,
                   keep_alive: CheckboxToolsAdvanced ? values.keepAliveTools : undefined,
-                }
-              : undefined,
-          embedding:
-            CheckboxEmbedding && values.serverEmbedding && values.modelEmbedding
-              ? {
-                  server_name: values.serverEmbedding,
-                  server: OllamaServer.get(values.serverEmbedding) as OllamaServer,
-                  tag: values.modelEmbedding,
-                  keep_alive: CheckboxEmbeddingAdvanced ? values.keepAliveEmbedding : undefined,
                 }
               : undefined,
         },
@@ -384,47 +335,6 @@ export function FormModel(props: props): JSX.Element {
       )}
       {CheckboxToolsAdvanced && (
         <Form.TextField title="Keep Alive" info={InfoKeepAlive} {...itemProps.keepAliveTools} />
-      )}
-      {!IsLoadingModel && Model && (
-        <React.Fragment>
-          <Form.Separator />
-          <Form.Checkbox
-            id="embedding"
-            info={InfoEmbeddingCheckbox}
-            title="Embedding"
-            label="Use Different Model"
-            defaultValue={CheckboxEmbedding}
-            onChange={SetCheckboxEmbedding}
-          />
-        </React.Fragment>
-      )}
-      {!IsLoadingModel && Model && itemProps && CheckboxEmbedding && (
-        <React.Fragment>
-          <Form.Dropdown title="Server" info={InfoServer} {...itemProps.serverEmbedding}>
-            {[...Model.keys()].sort().map((s) => (
-              <Form.Dropdown.Item title={s} value={s} key={s} />
-            ))}
-          </Form.Dropdown>
-          <Form.Dropdown title="Model" info={InfoModel} {...itemProps.modelEmbedding}>
-            {itemProps.serverEmbedding.value &&
-              Model.get(itemProps.serverEmbedding.value)!
-                .filter(
-                  (t) =>
-                    t.capabilities && t.capabilities.findIndex((c) => c === OllamaApiModelCapability.EMBEDDING) !== -1
-                )
-                .sort()
-                .map((s) => <Form.Dropdown.Item title={s.name} value={s.name} key={s.name} />)}
-          </Form.Dropdown>
-          <Form.Checkbox
-            id="advancedEmbedding"
-            label="Advanced Settings"
-            defaultValue={CheckboxEmbeddingAdvanced}
-            onChange={SetCheckboxEmbeddingAdvanced}
-          />
-        </React.Fragment>
-      )}
-      {CheckboxEmbeddingAdvanced && (
-        <Form.TextField title="Keep Alive" info={InfoKeepAlive} {...itemProps.keepAliveEmbedding} />
       )}
       {McpServer && !isLoadingMcpServer && (
         <React.Fragment>