+++ extensions/raycast-ollama/src/lib/ui/ChatView/main.tsx
@@ -8,12 +8,11 @@ import {
 } from "../../settings/settings";
 import { RaycastChat, RaycastChatMessage } from "../../settings/types";
 import { ChangeChat, ClipboardConversation, NewChat, Run } from "./function";
+import { Shortcut } from "../shortcut";
 import { FormModel } from "./form/Model";
 import { FormRenameChat } from "./form/RenameChat";
 import { GetImage } from "../function";
 import { RaycastImage } from "../../types";
-import { Document } from "langchain/document";
-import { FormAttachFile } from "./form/AttachFile";
 
 /**
  * Return JSX element for chat view.
@@ -41,10 +40,6 @@ export function ChatView(): JSX.Element {
     RaycastImage[] | undefined,
     React.Dispatch<React.SetStateAction<RaycastImage[] | undefined>>
   ] = React.useState();
-  const [Document, SetDocument]: [
-    Document<Record<string, any>>[] | undefined,
-    React.Dispatch<React.SetStateAction<Document<Record<string, any>>[] | undefined>>
-  ] = React.useState();
 
   // Save Chat To LocalStoarge on Inference Done.
   React.useEffect(() => {
@@ -52,7 +47,6 @@ export function ChatView(): JSX.Element {
       SetQuery("");
       SetIsLoading(false);
       if (Image) SetImage(undefined);
-      if (Document) SetDocument(undefined);
       if (Chat.messages.length === 1 && Chat.name === "New Chat")
         SetChat((prevValue) => {
           if (prevValue) {
@@ -88,9 +82,6 @@ export function ChatView(): JSX.Element {
   // Form: Model
   const [showFormModel, setShowFormModel]: [boolean, React.Dispatch<React.SetStateAction<boolean>>] =
     React.useState(false);
-  // Form: AttachDocument
-  const [showFormAttachFile, setShowFormAttachFile]: [boolean, React.Dispatch<React.SetStateAction<boolean>>] =
-    React.useState(false);
 
   /**
    * Action Panel for  Message
@@ -105,7 +96,7 @@ export function ChatView(): JSX.Element {
             title="Get Answer"
             icon={Icon.SpeechBubbleActive}
             onAction={() => {
-              Run(Query, Image, Document, Chat, SetChat, SetIsLoading).catch(async (e: Error) => {
+              Run(Query, Image, Chat, SetChat, SetIsLoading).catch(async (e: Error) => {
                 await showToast({ style: Toast.Style.Failure, title: "Error:", message: e.message });
                 SetIsLoading(false);
               });
@@ -124,14 +115,14 @@ export function ChatView(): JSX.Element {
             <Action.CopyToClipboard
               title="Copy Question"
               content={props.message.messages[0].content as string}
-              shortcut={{ modifiers: ["cmd"], key: "b" }}
+              shortcut={Shortcut.CopyName}
             />
           )}
           {props.message && (
             <Action.CopyToClipboard
               title="Copy Answer"
               content={props.message.messages[1].content as string}
-              shortcut={{ modifiers: ["cmd", "shift"], key: "c" }}
+              shortcut={Shortcut.Copy}
             />
           )}
           {props.message && <Action.CopyToClipboard title="Copy Conversation" content={ClipboardConversation(Chat)} />}
@@ -140,7 +131,7 @@ export function ChatView(): JSX.Element {
               title="New Chat"
               icon={Icon.NewDocument}
               onAction={() => NewChat(Chat, SetChatNameIndex, RevalidateChatNames)}
-              shortcut={{ modifiers: ["cmd"], key: "n" }}
+              shortcut={Shortcut.New}
             />
           )}
           {Chat && (
@@ -148,11 +139,11 @@ export function ChatView(): JSX.Element {
               title="Rename Chat"
               icon={Icon.Pencil}
               onAction={() => setShowFormRenameChat(true)}
-              shortcut={{ modifiers: ["cmd"], key: "e" }}
+              shortcut={Shortcut.Edit}
             />
           )}
           {Chat && (
-            <ActionPanel.Submenu title="Delete Chat" icon={Icon.Trash}>
+            <ActionPanel.Submenu title="Delete Chat" icon={Icon.Trash} shortcut={Shortcut.Remove}>
               <Action
                 title={`Yes, Delete "${Chat.name}" Chat`}
                 icon={Icon.Trash}
@@ -172,13 +163,13 @@ export function ChatView(): JSX.Element {
               title="Selection"
               icon={Icon.QuoteBlock}
               onAction={() => SetQuery((prevState) => (prevState += "\n{selection}\n"))}
-              shortcut={{ modifiers: ["cmd"], key: "s" }}
+              shortcut={Shortcut.AttachText}
             />
             <Action
               title="Browser Extention Tab"
               icon={Icon.Globe}
               onAction={() => SetQuery((prevState) => (prevState += "\n{browser-tab}\n"))}
-              shortcut={{ modifiers: ["cmd"], key: "b" }}
+              shortcut={Shortcut.AttachBrowserTab}
             />
             <Action
               title="Image From Clipboard"
@@ -193,13 +184,7 @@ export function ChatView(): JSX.Element {
                     showToast({ style: Toast.Style.Failure, title: "Error: ", message: String(e) });
                   })
               }
-              shortcut={{ modifiers: ["cmd"], key: "i" }}
-            />
-            <Action
-              title="File"
-              icon={Icon.Finder}
-              onAction={() => setShowFormAttachFile(true)}
-              shortcut={{ modifiers: ["cmd"], key: "f" }}
+              shortcut={Shortcut.AttachImage}
             />
           </ActionPanel.Section>
         )}
@@ -209,14 +194,14 @@ export function ChatView(): JSX.Element {
               title="Change Model"
               icon={Icon.Box}
               onAction={() => setShowFormModel(true)}
-              shortcut={{ modifiers: ["cmd"], key: "m" }}
+              shortcut={Shortcut.ChangeModel}
             />
           )}
           {props.message && (
             <Action
               title={ShowAnswerMetadata ? "Hide Metadata" : "Show Metadata"}
               icon={ShowAnswerMetadata ? Icon.EyeDisabled : Icon.Eye}
-              shortcut={{ modifiers: ["cmd"], key: "y" }}
+              shortcut={Shortcut.ToggleQuickLook}
               onAction={() => SetShowAnswerMetadata((prevState) => !prevState)}
             />
           )}
@@ -306,9 +291,6 @@ export function ChatView(): JSX.Element {
         revalidate={RevalidateChatNames}
       />
     );
-
-  if (showFormAttachFile)
-    return <FormAttachFile SetDocument={SetDocument} SetShow={setShowFormAttachFile} Document={Document} />;
 
   return (
     <List