+++ extensions/markdown-codeblock/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Markdown Codeblock Changelog
 
+## [Improvements] - 2025-11-21
+
+- Action titles now display the active application name instead of "Active App"
+- Action icons now display the active application icon
+
 ## [Improvements] - 2025-10-30
 
 - Add support for newline prefix before codeblocks